+++ src/components/modules/forms/SignUpForm.tsx
@@ -7,7 +7,7 @@ import { SignUpInit, SignUpRequest } from 'redux/reducers/SignUp';
 import { SignUpTypes } from 'types/module';
 import useInput from 'hooks/useInput';
 import useInputEnter from 'hooks/useInputEnter';
-import { formErrorHandler } from 'utils/formErrorHandler';
+import { passwordValidateHandler } from 'utils/validateCheckHandler';
 import { signUpErrorHandler } from 'utils/signUpErrorHandler';
 import DefaultButton from 'components/atoms/buttons/DefaultButton';
 import BorderInput from 'components/atoms/inputs/BorderInput';
@@ -27,7 +27,6 @@ function SignUpForm({ email, authDone, setAuthDone }: SignUpTypes) {
   const [passCompareError, setPassCompareError] = useState<boolean>(false);
   const [signUpError, setSignUpError] = useState<boolean>(false);
   const [registerDone, setRegisterDone] = useState<boolean>(true);
-  const { handleFormError } = formErrorHandler();
   const { handleSignUpError } = signUpErrorHandler();
   const { handleInputEnter } = useInputEnter();
   const signUpErrorStatus = useSelector(
@@ -35,7 +34,7 @@ function SignUpForm({ email, authDone, setAuthDone }: SignUpTypes) {
   );
 
   useEffect(() => {
-    handleFormError(
+    passwordValidateHandler(
       { username, nick, password, confirmPassword },
       {
         passError,
