+++ src/utils/validateCheckHandler.ts
@@ -1,5 +1,10 @@
-import { InputDataTypes, ValidateCheckTypes } from 'types/util';
-import { codeRegex, emailRegex } from 'constants/regex';
+import {
+  InputDataTypes,
+  ValidateCheckTypes,
+  CompareErrorTypes,
+  SignUpDataTypes,
+} from 'types/util';
+import { codeRegex, emailRegex, passwordRegex } from 'constants/regex';
 
 export const validateCheckHandler = (
   values: InputDataTypes,
@@ -15,5 +20,30 @@ export const validateCheckHandler = (
     validateCheck.setCodeValidate(false);
   } else if (values.code && codeRegex.test(values.code)) {
     validateCheck.setCodeValidate(true);
+  }
+};
+
+export const passwordValidateHandler = (
+  values: SignUpDataTypes,
+  error: CompareErrorTypes,
+) => {
+  if (values.password && !passwordRegex.test(values.password)) {
+    error.setPassError(true);
+  } else if (values.password && passwordRegex.test(values.password)) {
+    error.setPassError(false);
+  }
+
+  if (
+    !error.passError &&
+    values.confirmPassword &&
+    values.password !== values.confirmPassword
+  ) {
+    error.setPassCompareError(true);
+  } else if (
+    !error.passError &&
+    values.confirmPassword &&
+    values.password === values.confirmPassword
+  ) {
+    error.setPassCompareError(false);
   }
 };
